refactor(docs): split up util functions
