+++ docs/app/utils/parentComponents.js
@@ -1,0 +1,8 @@
+import _ from 'lodash/fp'
+
+import * as semanticUIReact from 'src'
+import { META } from 'src/lib'
+
+const parentComponents = _.flow(_.filter(META.isParent), _.sortBy('_meta.name'))(semanticUIReact)
+
+export default parentComponents
